fixed topic change
The code of the topic module used to work unreliably.
I hope to have fixed this with my changes.

First, the channel names are now always normalized
to lowercase letters. Second, the bot explicitly
updates its knowledge of the channel's topic, which
I think hasn't been done before, causing the bot to
"forget" the topic from time to time ...