Hide the RNG state via the global IO monad RNG
We use the hidden RNG inside the IO monad to achieve random number
generation for lambdabot. This cleans up the code, since we do not
need to have explicit passing of the RNG state in the functions and
we furthermore do not need many of the functions the net sum of this
change is -55 lines of code, with no functional change.