prevent static modules from being loaded and unloaded, experimental.
This is one way to solve the problem that admins can unload static modules,
which aren't accessible anymore after that. The patch changes 
GenStaticModules.hs to generate a second module StaticModules2.hs which
exports a function that checks if a name belongs to a static module.

I'm not really happy with this patch, but I didn't find a simpler solution
and I didn't want to add additional state to the bot (maybe that's a better
way to go?)

Note: When run, lambdabot complains about an undefined symbol for no apperent 
reason.