handle persistance centrally.
The patch allows modules which keep state to easily save and restore it by
implementing moduleSerialize which returns a "Serializer" (probably needs a
better name) that can transform the state to a String and back. Currently, the
state is only written upon exit, but that seems a little dangerous.
Alternatives:
* Flush on every writeMS to a watched state.
* Provide an explicit write command ? la @eval's dump (which has been removed).
* Any combination of the above possibilities.

Minor changes:
* commands has been renamed to the more consistent moduleCmds.
* the definition of theModule has been shrinked in a most (?) modules.
* moduleName is pure now, but maybe it should be dropped alltogether.
* readM has been added to Utils to replace calls of read that may fail.