Replace ReaderT by implicit parameters.
On #haskell, most people didn't like the idea of implicit parameters, but in 
our case it simplifies things quite a bit, because one doesn't need to cope
with ReaderT transformer, if the module doesn't need state. Also, forall-
hoisting eliminates the need for ircConnectSignalR quite nicely.

So it's up to you to trade between the (imho, but others disagree) mild
evilness of implicit parameters and the slighly more complicated implementation
of modules (lifting, mapReaderT liftLB, etc.).

It's a small change anyway.