Revert to the old behavior, i.e don't send NOTICEs anymore.
Justification:
~~~~~~~~~~~~~~

Many IRC clients don't interpret NOTICES properly (i.e. as mandated by the
RFC), annoying their users. They fall into two categories, those which can
easily be tweaked to behave sanely (at least X-Chat, irssi) and those which
cannot (e.g. gaim). I agree that these clients SHOULD be fixed, but let's 
not fight that war on the back of #haskell's users. It is inacceptable to
tell them to switch their client (might even be impossible, I recall someone
using BeOS complaining, who knows how many IRC clients there are for that
platform).
That is not to say, we shouldn't reconsider the decision later (though
I very much doubt that most clients will be compliant in the near future).

The looping problem remains. We have to watch out for the following cases:
1) Two current lambdabots are in the channel.
2) Our lambdabot and an old version are present.
3) Lambdabot + another bot.
4) A determined attacker is trying to annoy the rest of the channel.

1) and 2) are solved by the workaround this patch implements: Simply "escape"
all commands starting with a '@' with a space. In case 3), loops might still
be possible, but very unlikely, in my opinion. Nothing can reasonably protect
from 4), "/ban" is the proper way to deal with that problem.