HEADS UP: use hs-plugins for dynamic loading
We now use hs-plugins to dynamically load the whole bot, from a static
core defined in Boot.hs. DynamicModule now uses hs-plugins functions,
which it accesses through Boot.hs. We no longer statically and
dynamically link the bot core, which is good.

Requires hs-plugins 0.9.8 be installed and registered.

We shoule be able to do some fancy things with this.