StateT s IO --> ReaderT (IORef s) IO, remove ExceptionErrorT.
This patch removes some unnecessary layers of indirection, which hopefully
makes the code less obscure. Furthermore, it lifts the restriction that the
state must be accessed in a single-threaded way. I'm also thinking of the
possibility of a dynamic LB -> IRC cast, which would allow modules to send
messages even when their 'process' method isn't invoked by the event handler.

This is an experimental change, but it should behave exactly like the old 
version (that means, some of the problems remain).