added plugin Pretty, for pretty printed code echoing
this plugin is supposed to help in explaining language features to newbies, which can at times be difficult on a single line interface.
the code given to the command @pretty will be parsed by Language.Haskell.Parser and pretty-printed/indented by Language.Haskell.Pretty. any code, be it a let-expression, a single declaration or multiple declarations, is supported.