I'd fallen foul of lambdabot's excessive output more than once. A simpl e omission of a "take" when dealing with infinite lists can spam the ch annel and annoy everyone watching.
This is a simple patch to cut off lambdabot's output to 80 characters. If the output was going to be more than that, the first 77 and "..." are printed.

We could perhaps change the length of output based on whether or not we're in a /query (or /msg; whatever) with someone, or are speaking to a channel.
Only output 80 chars