Refactored the URL page and title fetching to a URL library.
The intent of this is to allow other plugins to leverage the fetching of
pages (and following of redirects) as well as the fetching of page
titles for text/html content.  It was suggested by dmhouse that I modify
the Search plugin to print not only the URL of the result, but also the
title of it.