Replace ircStayConnected with ircPersists
Now we have a map of servers to which we try to reconnect, including
offlinerc. Reconnection can be one-time (reconnect once) or lasting (reconnect
always). We use new data structure for offlinerc instead of
ircStayConnected -- as a side effect, this fixes a bug when `quit`ting from any
IRC server also stops offlinerc.